move App components to HomePage and import HomePage
