+++ src/App.tsx
@@ -1,17 +1,9 @@
 import './App.module.scss';
 import React, { useState, useEffect } from 'react';
 import { Loader } from './components/Loader/Loader';
-import { BannerSlider } from './components/BannerSlider';
-import { Header } from './components/Header';
-import { Footer } from './components/Footer';
-import { ProductSlider } from './components/ProductSlider';
-import { MobileSwiper } from './components/Swiper';
-import { Thumbnails } from './components/Thumbnails';
-import { PreviousPage } from './components/PreviousPage';
-import { useLocation } from 'react-router-dom';
+import { HomePage } from './modules/HomePage'
 
 export const App: React.FC = () => {
-  const category = useLocation().pathname.slice(1)
   const [isLoading, setIsLoading] = useState(true);
   useEffect(() => {
     setIsLoading(false);
@@ -23,14 +15,7 @@ export const App: React.FC = () => {
         <Loader />
       ) : (
         <div>
-          <PreviousPage category= {category}/>
-          <Header />
-          <BannerSlider />
-          <MobileSwiper />
-          <ProductSlider title="Brand new models" sortMethod={'newest'} category={'phones'}/>
-          <Thumbnails />
-          <ProductSlider title="Hot prices" sortMethod={'hot'} category={'phones'}/>
-          <Footer />
+          <HomePage />
         </div>
       )};
     </div>
